Add function to generate choice randomly
In this version of the rock, paper, scissors game, the player goes
against the computer; in order to generate a random choice for the
computer there are two functions: randomIntRange() and computerPlay().

randomIntRange() has two parameters that define a range and then returns
a value between that range using a mathematical expression.

computerPlay() is the function that returns a string with the choice.
It has a variable that stores a random number between 1 and 3 and based
on some conditionals it returns the choice.