Add game() function
The game() function is a match between the player and the computer of
five rounds and reports the winner of the majority or a tie if both won
the same amount of rounds.

It uses a for loop to call five times the playRound() function and
stores the results in two variables, one for each player.

At the end of the loop, it determines the winner or the tie based on
the results of the games and displays a message on the console.