Update the score of the players
With some if conditions in the button's function it's called the new
updateScore() Function.

This function receives the class of the element
that won the round when a button is clicked. Then it creates a copy
of the element, creates an array to obtain the score, then converts
that string value to an int in order to increase in 1 the value and
finally update the content of the element to display the result on
the page.